rename options to settings
