--- src/Apps/Windows/AnyStatus.Apps.Windows/Features/Options/OptionsView.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace AnyStatus.Apps.Windows.Features.Options
-{
-    public partial class OptionsView : UserControl
-    {
-        public OptionsView()
-        {
-            InitializeComponent();
-        }
-    }
-}
