+++ src/Apps/Windows/AnyStatus.Apps.Windows/Features/Settings/SettingsView.xaml
@@ -1,12 +1,12 @@
-﻿<UserControl x:Class="AnyStatus.Apps.Windows.Features.Options.OptionsView"
+﻿<UserControl x:Class="AnyStatus.Apps.Windows.Features.Settings.SettingsView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:AnyStatus.Apps.Windows.Features.Options"
+             xmlns:local="clr-namespace:AnyStatus.Apps.Windows.Features.Settings"
              xmlns:propertygrid="clr-namespace:AnyStatus.Apps.Windows.Infrastructure.Mvvm.Controls.PropertyGrid"
              mc:Ignorable="d"
-             d:DataContext="{d:DesignInstance Type=local:OptionsViewModel, IsDesignTimeCreatable=False}">
+             d:DataContext="{d:DesignInstance Type=local:SettingsViewModel, IsDesignTimeCreatable=False}">
     <ScrollViewer>
         <StackPanel Margin="10">
             <propertygrid:PropertyGridControl DataContext="{Binding PropertyGridViewModel}"/>
