+++ src/Apps/Windows/AnyStatus.Apps.Windows/Features/Settings/SettingsView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Controls;
+
+namespace AnyStatus.Apps.Windows.Features.Settings
+{
+    public partial class SettingsView : UserControl
+    {
+        public SettingsView()
+        {
+            InitializeComponent();
+        }
+    }
+}
